Online transportation booking/ based on Ethiopia
Online Public Transport Booking This project is based on a client-server architecture. It is a full-stack MERN project which contains two react-native applications (for a driver and a user) and a two react based web applications (for an admin and an agent). Server side application is developed using MongoDB, ExpressJS and NodeJS. Client side application: The User side application allows the individual to: - signup and login, - enter their current location and their destination location, - choose a mode of transport, be able to see price of route, and make payment - receive a generatedb unique QR code as a ticket, - add comments regarding the service. The Driver side application allows the individual to: - signup and login, - be able to recieve travel schedules and view route, - scan a user's QR code ticket to make sure it's a valid ticket, - be able to contact an agent for help in cases of traffic jam or other incidents, - add comments regarding the service. The Admin side web application allows the individual to: - signup and login, - view list of user's who use the system, - view list of driver's who use the system. The Agent side web application allows the individual to: - signup and login, - manage routes to assist the driver, - review comments from user's and drivers.